% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecart_type.R, R/Loi_normale.R
\name{ecart_type}
\alias{ecart_type}
\title{Définir une fonction qui calcule l'écart-type empirique d'un vecteur}
\usage{
ecart_type(x)

ecart_type(x)
}
\arguments{
\item{x}{un vecteur numérique}
}
\value{
un nombre réel qui représente l'écart-type empirique

un nombre réel qui représente l'écart-type empirique
}
\description{
Définir une fonction qui calcule l'écart-type empirique d'un vecteur

Définir une fonction qui calcule l'écart-type empirique d'un vecteur
}
