% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecart_type.R, R/Echantillon.R, R/Loi_normale.R,
%   R/Moyenne.R, R/Variance.R
\name{echantillon}
\alias{echantillon}
\title{Définir une fonction qui génère un échantillon aléatoire d'une distribution donnée
avec une taille et des paramètres spécifiés}
\usage{
echantillon(dist, size, ...)

echantillon(dist, size, ...)

echantillon(dist, size, ...)

echantillon(dist, size, ...)

echantillon(dist, size, ...)
}
\arguments{
\item{dist}{une chaîne de caractères qui indique le nom de la distribution}

\item{size}{un entier positif qui indique la taille de l'échantillon}

\item{...}{les paramètres de la distribution}
}
\value{
un vecteur numérique qui contient l'échantillon aléatoire

un vecteur numérique qui contient l'échantillon aléatoire

un vecteur numérique qui contient l'échantillon aléatoire

un vecteur numérique qui contient l'échantillon aléatoire

un vecteur numérique qui contient l'échantillon aléatoire
}
\description{
Définir une fonction qui génère un échantillon aléatoire d'une distribution donnée
avec une taille et des paramètres spécifiés

Définir une fonction qui génère un échantillon aléatoire d'une distribution donnée
avec une taille et des paramètres spécifiés

Définir une fonction qui génère un échantillon aléatoire d'une distribution donnée
avec une taille et des paramètres spécifiés

Définir une fonction qui génère un échantillon aléatoire d'une distribution donnée
avec une taille et des paramètres spécifiés

Définir une fonction qui génère un échantillon aléatoire d'une distribution donnée
avec une taille et des paramètres spécifiés
}
