% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loi_normale.R
\name{loi_normale}
\alias{loi_normale}
\title{Définir une fonction qui calcule la loi normale d'une valeur x
en fonction de la moyenne mu et de l'écart-type sigma}
\usage{
loi_normale(x, mu, sigma)
}
\arguments{
\item{x}{un nombre réel}

\item{mu}{un nombre réel qui représente la moyenne}

\item{sigma}{un nombre réel positif qui représente l'écart-type}
}
\value{
un nombre réel qui représente la valeur de la loi normale
}
\description{
Définir une fonction qui calcule la loi normale d'une valeur x
en fonction de la moyenne mu et de l'écart-type sigma
}
