% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecart_type.R, R/Loi_normale.R, R/Moyenne.R,
%   R/Variance.R
\name{moyenne}
\alias{moyenne}
\title{Définir une fonction qui calcule la moyenne empirique d'un vecteur}
\usage{
moyenne(x)

moyenne(x)

moyenne(x)

moyenne(x)
}
\arguments{
\item{x}{un vecteur numérique}
}
\value{
un nombre réel qui représente la moyenne empirique

un nombre réel qui représente la moyenne empirique

un nombre réel qui représente la moyenne empirique

un nombre réel qui représente la moyenne empirique
}
\description{
Définir une fonction qui calcule la moyenne empirique d'un vecteur

Définir une fonction qui calcule la moyenne empirique d'un vecteur

Définir une fonction qui calcule la moyenne empirique d'un vecteur

Définir une fonction qui calcule la moyenne empirique d'un vecteur
}
