% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoreme_central_limite.R
\name{theoreme_central_limite}
\alias{theoreme_central_limite}
\title{Définir une fonction qui applique le théorème central limite à un échantillon
en calculant la distribution de la moyenne empirique}
\usage{
theoreme_central_limite(x, n, m)
}
\arguments{
\item{x}{un vecteur numérique qui représente l'échantillon}

\item{n}{un entier positif qui représente le nombre de sous-échantillons à tirer}

\item{m}{un entier positif qui représente la taille des sous-échantillons}
}
\value{
une liste qui contient les éléments suivants:
\itemize{
\item moyennes: un vecteur numérique qui contient les moyennes empiriques des sous-échantillons
\item mu: un nombre réel qui représente la moyenne théorique de la distribution de la moyenne empirique
\item sigma: un nombre réel qui représente l'écart-type théorique de la distribution de la moyenne empirique
\item densite: une fonction qui calcule la densité de la distribution de la moyenne empirique
}
}
\description{
Définir une fonction qui applique le théorème central limite à un échantillon
en calculant la distribution de la moyenne empirique
}
\examples{
echantillon_normale <- echantillon("normale", 1000, mean = 10, sd = 2)
#Appliquer le théorème central limite à cet échantillon avec n = 100 et m = 50
resultat <- theoreme_central_limite(echantillon_normale, 100, 50)
#Afficher les résultats
print(paste("La moyenne théorique de la distribution de la moyenne empirique est", resultat$mu))
print(paste("L'écart-type théorique de la distribution de la moyenne empirique est", resultat$sigma))
hist(resultat$moyennes, freq = FALSE, main = "Histogramme des moyennes empiriques", xlab = "Moyenne empirique")+
curve(resultat$densite(x), add = TRUE, col = "red", lwd = 2)

}
