% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecart_type.R, R/Loi_normale.R, R/Variance.R
\name{variance}
\alias{variance}
\title{Définir une fonction qui calcule la variance empirique d'un vecteur}
\usage{
variance(x)

variance(x)

variance(x)
}
\arguments{
\item{x}{un vecteur numérique}
}
\value{
un nombre réel qui représente la variance empirique

un nombre réel qui représente la variance empirique

un nombre réel qui représente la variance empirique
}
\description{
Définir une fonction qui calcule la variance empirique d'un vecteur

Définir une fonction qui calcule la variance empirique d'un vecteur

Définir une fonction qui calcule la variance empirique d'un vecteur
}
